+++ vagrant/setup.sh
@@ -1,7 +1,7 @@
 #!/bin/bash
 
 # Install pip and python development libs
-apt-get -y install python-pip python-dev
+apt-get -y install python-pip python-dev libmysqlclient-dev
 # Install pythonb libs
 pip install -r /vagrant/requirements.txt
 # Make python-oauth2 available for python
@@ -9,5 +9,7 @@ if ! grep -Fxq "export PYTHONPATH=/opt/python-oauth2" /home/vagrant/.bashrc
 then
     echo "export PYTHONPATH=/opt/python-oauth2" >> /home/vagrant/.bashrc
 fi
+# Create the testdb database in mysql
+mysql -uroot -pmaster < /vagrant/mysql-schema.sql
 # Execute script to create a testclient entry in mongodb
 python /vagrant/create_testclient.py